pyproject.toml
[project]
name = "computer-vision-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "joblib>=1.5.3",
    "mediapipe>=0.10.32",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "python-fasthtml>=0.12.50",
    "starlette>=0.52.1",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
